Minor changes to get windows running simple strings but failing on loading certain modules. Looks like modules on win32 are dlls and for some reason not understood that means perhaps it must be linked differently than worked on linux... maybe?

+++ common/quaxol.cpp
@@ -304,7 +304,7 @@ void QuaxolChunk::AddRenderCubeByFlag(
   CanonicalCube& cube = s_canonicalCubesByFlag[connectFlags];
   assert(cube.m_connectFlags == connectFlags); // call BuildCanonicalCubesByFlag
 
-  int indexOffset = m_verts.size();
+  int indexOffset = (int)m_verts.size();
   for(auto vert : cube.m_verts) {
     m_verts.emplace_back(vert + vertOffset);
   }
@@ -324,7 +324,7 @@ void QuaxolChunk::AddRenderCubeByDir(
   CanonicalCube& cube = s_canonicalCubesByDir[dirIndex];
   assert(cube.m_dirIndex == dirIndex); // call BuildCanonicalCubesByDir
 
-  int indexOffset = m_verts.size();
+  int indexOffset = (int)m_verts.size();
   for(auto vert : cube.m_verts) {
     m_verts.emplace_back(vert + vertOffset);
   }
